refactor config database slightly
isolated each type of db to limit the impact of dependencies when needing to install on a device. also changed the places that used it to have an instance of one passed in so it doesn't need to know how they work
