+++ configure_devices_test.py
@@ -1,11 +1,10 @@
-import decawave_ble
 import configure_devices
+from config.csv import ConfigurationDatabaseCSVLocal
+
 
 # Configuration databases
 config_database_path = 'config_data/tech_room_config_database.csv'
 
 # Configure from database
 print('\nConfiguring from database')
-configure_devices.configure_devices_from_database(config_database_path)
-
-
+configure_devices.configure_devices_from_database(ConfigurationDatabaseCSVLocal(config_database_path))
